+++ extensions/sportssync/src/f1-standings.tsx
@@ -1,4 +1,4 @@
-import { Detail, List } from "@raycast/api";
+import { Detail, List, Icon, Color } from "@raycast/api";
 import { useFetch } from "@raycast/utils";
 import { useState } from "react";
 
@@ -51,12 +51,32 @@ export default function scoresAndSchedule() {
   const drivers = driverItems.map((driver, index) => {
     const flagSrc = driver?.athlete?.flag.href ?? `${driver?.athlete?.flag?.href}`;
 
+    const driverPosition = Number(driver.stats[0].displayValue);
+
+    let tagColor;
+    let tagIcon;
+
+    if (driverPosition === 1) {
+      tagColor = Color.Yellow;
+      tagIcon = Icon.Trophy;
+    } else if (driverPosition >= 2) {
+      tagColor = Color.Green;
+      tagIcon = Icon.Leaderboard;
+    } else {
+      tagColor = Color.SecondaryText;
+    }
+
     return (
       <List.Item
         key={index}
         title={`${driver?.athlete?.displayName}`}
         icon={{ source: flagSrc }}
-        accessoryTitle={`${driver.stats[1].displayValue} pts `}
+        accessories={[
+          {
+            text: `${driver.stats[1].displayValue} pts`,
+          },
+          { tag: { value: driver.stats[0].displayValue, color: tagColor }, icon: tagIcon },
+        ]}
       />
     );
   });
@@ -69,11 +89,31 @@ export default function scoresAndSchedule() {
 
   const constructorItems = constructorData?.children?.[1]?.standings?.entries || [];
   const constructorTeams = constructorItems.map((constructor, index) => {
+    const constructorPosition = Number(constructor.stats[0].displayValue);
+
+    let tagColor;
+    let tagIcon;
+
+    if (constructorPosition === 1) {
+      tagColor = Color.Yellow;
+      tagIcon = Icon.Trophy;
+    } else if (constructorPosition >= 2) {
+      tagColor = Color.Green;
+      tagIcon = Icon.Leaderboard;
+    } else {
+      tagColor = Color.SecondaryText;
+    }
+
     return (
       <List.Item
         key={index}
         title={`${constructor?.team?.displayName}`}
-        accessoryTitle={`${constructor.stats[1].displayValue} pts `}
+        accessories={[
+          {
+            text: `${constructor.stats[1].displayValue} pts`,
+          },
+          { tag: { value: constructor.stats[0].displayValue, color: tagColor }, icon: tagIcon },
+        ]}
       />
     );
   });