+++ extensions/sportssync/src/mlb-scores-and-schedule.tsx
@@ -1,4 +1,4 @@
-import { Detail, List, Color, Action, ActionPanel } from "@raycast/api";
+import { Detail, List, Color, Icon, Action, ActionPanel } from "@raycast/api";
 import { useFetch } from "@raycast/utils";
 import getPastAndFutureDays from "./utils/getDateRange";
 
@@ -15,9 +15,8 @@ interface Status {
   type: {
     state: string;
     completed?: boolean;
+    detail?: string;
   };
-  period?: number;
-  displayClock?: string;
 }
 
 interface Competition {
@@ -76,34 +75,26 @@ export default function command() {
 
     let accessoryTitle = gameTime;
     let accessoryColor = Color.SecondaryText;
+    let accessoryIcon = { source: Icon.Calendar, tintColor: Color.SecondaryText };
     let accessoryToolTip;
 
-    function getPeriodWithSuffix(period: number | undefined): string {
-      if (period === undefined) return "";
-      if (period === 1) return `${period}st`;
-      if (period === 2) return `${period}nd`;
-      if (period === 3) return `${period}rd`;
-      if (period >= 4 && period <= 9) return `${period}th`;
-      return `${period}`;
-    }
-
-    const period = game.status.period;
-    const periodWithSuffix = getPeriodWithSuffix(period);
-
     if (game.status.type.state === "in") {
-      accessoryTitle = `${game.competitions[0].competitors[1].team.abbreviation} ${game.competitions[0].competitors[1].score} - ${game.competitions[0].competitors[0].team.abbreviation} ${game.competitions[0].competitors[0].score}     ${periodWithSuffix} ${game.status.displayClock}`;
+      accessoryTitle = `${game.competitions[0].competitors[1].team.abbreviation} ${game.competitions[0].competitors[1].score} - ${game.competitions[0].competitors[0].team.abbreviation} ${game.competitions[0].competitors[0].score}     ${game.status.type.detail}`;
       accessoryColor = Color.Green;
+      accessoryIcon = { source: Icon.Livestream, tintColor: Color.Green };
       accessoryToolTip = "In Progress";
     }
 
     if (game.status.type.state === "post") {
       accessoryTitle = `${game.competitions[0].competitors[1].team.abbreviation} ${game.competitions[0].competitors[1].score} - ${game.competitions[0].competitors[0].team.abbreviation} ${game.competitions[0].competitors[0].score}`;
       accessoryColor = Color.SecondaryText;
+      accessoryIcon = { source: Icon.CheckCircle, tintColor: Color.SecondaryText };
       accessoryToolTip = "Final";
     }
 
     if (game.status.type.state === "post" && game.status.type.completed === false) {
       accessoryTitle = `Postponed`;
+      accessoryIcon = { source: Icon.XMarkCircle, tintColor: Color.Orange };
       accessoryColor = Color.Orange;
     }
 
@@ -112,7 +103,10 @@ export default function command() {
         key={index}
         title={`${game.name}`}
         icon={{ source: game.competitions[0].competitors[1].team.logo }}
-        accessories={[{ text: { value: `${accessoryTitle}`, color: accessoryColor }, tooltip: accessoryToolTip }]}
+        accessories={[
+          { text: { value: `${accessoryTitle}`, color: accessoryColor }, tooltip: accessoryToolTip },
+          { icon: accessoryIcon },
+        ]}
         actions={
           <ActionPanel>
             <Action.OpenInBrowser title="View Game Details on ESPN" url={`${game.links[0].href}`} />