+++ extensions/sportssync/src/soccer-news.tsx
@@ -1,9 +1,10 @@
-import { Detail, List, Action, ActionPanel } from "@raycast/api";
+import { Detail, List, Action, ActionPanel, Color, Icon } from "@raycast/api";
 import { useFetch } from "@raycast/utils";
 
 interface Article {
   headline: string;
   published: string;
+  type: string;
   images: { url: string }[];
   links: { web: { href: string } };
 }
@@ -29,6 +30,11 @@ export default function scoresAndSchedule() {
 
     const accessoryTitle = articleDate;
     const accessoryToolTip = "Date Published";
+    let articleType = soccerArticle.type;
+
+    if (articleType === "HeadlineNews") {
+      articleType = "Headline";
+    }
 
     return (
       <List.Item
@@ -37,7 +43,11 @@ export default function scoresAndSchedule() {
         icon={
           soccerArticle.images && soccerArticle.images[0]?.url ? { source: soccerArticle.images[0].url } : undefined
         }
-        accessories={[{ text: { value: `${accessoryTitle}` }, tooltip: accessoryToolTip }]}
+        accessories={[
+          { tag: { value: articleType, color: Color.Green }, icon: Icon.Megaphone },
+          { text: { value: `${accessoryTitle}` }, tooltip: accessoryToolTip },
+          { icon: Icon.Calendar },
+        ]}
         actions={
           <ActionPanel>
             <Action.OpenInBrowser title="View Article on ESPN" url={`${soccerArticle.links.web.href}`} />